requirements-free.txt
fastapi
uvicorn
python-telegram-bot
sqlmodel[sqlite]
python-dotenv
nltk
sumy